Duke3D: Process q16rotscrnang and q16look_ang at frame-rate.
Had to make lastInputTicks a global so it could be reset in P_ResetPlayer(). When first running P_GetInput(), the initial value of elapsedInputTicks is the actual value of timerGetHiTicks(), which is into the thousands. This high initial value was affecting how scaleAdjustmentToInterval() scales as it was taking an initial q16look_ang value of 512 and overflowing the fix16_t type, making it -32,768.