Duke3D: Add CON command `setmusicvolume <percentage>`
This command is intended to allow scripts to temporarily lower the music volume.
It sets the current volume to `((<percentage> * ud.config.MusicVolume) / 100)`
while preserving the user's chosen volume defined in the sound settings menu.

`<percentage>` is hereby a value between 0 and 100 (inclusive).

For example, call `setmusicvolume 50` to reduce the music volume to 50%
of the user's chosen volume, and `setmusicvolume 100` to reset it.