Duke3d: tweak default JoystickAimWeight and JoystickViewCentering settings
I could make something up about how the new dead zone implementation makes the weighted aiming less necessary, but I really just wanted the defaults to be in the middle of the slidebar. The new dead zone implementation really is better enough that the weighted aiming is less necessary, though.