Duke3D: Allow undefining a skill without undefining everything above that level.
For example, it is now possible to undefine all skills except "Come Get Some".
Furthermore, you can now undefine all skills, in which case the menu is
skipped, and the default skill level is used. This was always the intended
behavior, but crashed the game due to a bug.

Fixes terminx/eduke32#95 and closes terminx/eduke32#96.