Duke3D: Perform soft reboot when launching addons from main menu
This is a very rough first draft to the process.

In G_SoftReboot() the game's state is supposed to be reset to the state
it was in just after the startup menu was closed. This requires resetting
a large number of structs, as well as resetting and initializing the
correct paths, GRPs, CONs, DEFs and RTS files.

The process currently works for GAME.CON addons and ART replacements.
Known issues:
* Repeated loading of addons crashes the game.
* Mouse input no longer works after rebooting.

Misc changes:
* Fix loading of addon script file paths.
* Some improvements for DEF commands, including newgamechoices.
* Create a read-only array for cheat name and description defaults.