Duke3D: Add commands "setrelativemusicvolume" and "getrelativemusicvolume"
Valid values are between 0 and 256 (inclusive).

This command allows mods to reduce the volume of the currently playing
music relative to the volume defined by the player. A modifier of 128 is
half the volume defined by the player, for instance.

The command "getrelativemusicvolume" stores the current modifier value
in the provided variable.

The modifier is set to 256 on game start, but is not reset afterwards.
In other words -- restarting a level, changing music tracks within a
level, resetting the sound system and music volume changes by the player
will all retain the currently defined volume reduction.

This means that the modifier needs to be controlled entirely through CON
code. It is up to the modder to ensure that the modifier is reset properly.