SW: Fix multiplayer sync errors with the weapon auto-switching toggle
This also permanently enables weapon auto-switching for PedanticMode.
The idea here is that outside of pedantic mode, within multiplayer
games, a new bit or more of the inspected player's Flags2 field will
be set for the weapon. Afterwards, only for player no. myconnectindex,
the function getinput will set loc->bits as if a matching key was
pressed for selecting the weapon. loc->bits should be properly synced.

Note that each player's structure, including Flags2, is reset
within InitNewGame, but not from InitLevel as far as I may tell.