SW: Add template function TerminateWithMsg and wrapper TerminateWithSimpleMsg
TerminateWithMsg is going to be used instead of separate
sets of calls to TerminateGame, printf/wm_msgbox and exit.
It'll use wm_msgbox, not printf (originally used under DOS),
and also go through LOG_F. This is an inline function
definition using a variadic template, allowing the
use of LOG_F and also matching loguru::log.

TerminateWithSimpleMsg is a wrapper that allows
printing a string without formatting.